Extract useClaimableRewards into a separate hook.
